[Chore] Refactor to MVP [Model-View-Presenter] Architecture Pattern
