+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@ package="com.example.antonyng.level_up">
     android:supportsRtl="true"
     android:theme="@style/AppTheme"
     android:configChanges="orientation|keyboardHidden">
-    <activity android:name=".view.MainActivity">
+    <activity android:name=".home.home.view.MainActivity">
         <intent-filter>
             <action android:name="android.intent.action.MAIN" />
 
@@ -20,8 +20,8 @@ package="com.example.antonyng.level_up">
         </intent-filter>
     </activity>
     <activity
-        android:name=".view.DetailActivity"
-        android:parentActivityName=".view.MainActivity" />
+        android:name=".github_user_profile.view.DetailActivity"
+        android:parentActivityName=".home.home.view.MainActivity" />
 </application>
 
 </manifest>
