+++ app/src/main/java/com/example/antonyng/level_up/home/home/contract/MainContract.java
@@ -1,6 +1,6 @@
-package com.example.antonyng.level_up.contract;
+package com.example.antonyng.level_up.home.home.contract;
 
-import com.example.antonyng.level_up.model.GithubUsers;
+import com.example.antonyng.level_up.home.home.model.GithubUsers;
 
 import java.util.ArrayList;
