+++ app/src/main/java/com/example/antonyng/level_up/home/home/model/GithubUsersResponse.java
@@ -1,9 +1,8 @@
-package com.example.antonyng.level_up.model;
+package com.example.antonyng.level_up.home.home.model;
 
 import com.google.gson.annotations.SerializedName;
 
 import java.util.ArrayList;
-import java.util.List;
 
 public class GithubUsersResponse {
     @SerializedName("items")
