+++ app/src/main/java/com/example/antonyng/level_up/home/home/service/GithubApi.java
@@ -1,6 +1,6 @@
-package com.example.antonyng.level_up.service;
+package com.example.antonyng.level_up.home.home.service;
 
-import com.example.antonyng.level_up.model.GithubUsersResponse;
+import com.example.antonyng.level_up.home.home.model.GithubUsersResponse;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
