+++ app/src/main/res/layout/content_main.xml
@@ -7,7 +7,7 @@
     android:id="@+id/swipe_container"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:showIn="@layout/activity_main"
-    tools:context=".view.MainActivity"
+    tools:context=".home.home.view.MainActivity"
     app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
     <android.support.v7.widget.RecyclerView
